Test

See merge request s07-final/S07P31C208!30

+++ src/main/java/com/ssafy/BravoSilverLife/entity/BDCode.java
@@ -14,8 +14,9 @@ import javax.persistence.Table;
 @NoArgsConstructor
 @Table(name = "bd_code")
 public class BDCode {
+
     @Id
-    int code;
+    long code;
     String sido;
     String gugun;
     String name;
